refactor: CodeGenerator를 추상화하고 redis 관련 CodeGenerator를 모두 제거한다 --- refactor: CodeGenerator 추상화 refactor: UuidCodeGenerator 구현체 생성 refactor: Redis 관련 CodeGenerator 모두 제거 refactor: Redis 관련 테스트 모두 제거
