--- src/main/java/com/fourweekdays/fourweekdays/common/generator/service/PessimisticLockService.java
@@ -1,4 +1,0 @@
-package com.fourweekdays.fourweekdays.common.generator.service;
-
-public class PessimisticLockService {
-}
