--- src/test/java/com/fourweekdays/fourweekdays/redis/InventoryTestV1.java
@@ -1,51 +1,0 @@
-package com.fourweekdays.fourweekdays.redis;
-
-import com.fourweekdays.fourweekdays.common.generator.CodeGenerator;
-import com.fourweekdays.fourweekdays.inbound.repository.InboundRepository;
-import com.fourweekdays.fourweekdays.inventory.repository.InventoryRepository;
-import com.fourweekdays.fourweekdays.inventory.service.InventoryService;
-import com.fourweekdays.fourweekdays.location.repository.LocationRepository;
-import com.fourweekdays.fourweekdays.product.repository.ProductRepository;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-@SpringBootTest
-public class InventoryTestV1 {
-
-    @Autowired
-    private InventoryService inventoryService;
-
-    @Test
-    void 재고증가() throws Exception {
-        Long inboundId = 1L;
-        String locationCode = "LOC-001";
-
-        int threadCount = 100;
-        ExecutorService executor = Executors.newFixedThreadPool(threadCount);
-        CountDownLatch latch = new CountDownLatch(threadCount);
-
-        for (int i = 0; i < threadCount; i++) {
-            executor.execute(() -> {
-                try {
-                    inventoryService.createInventoryFromInbound(inboundId, locationCode);
-                } catch (Exception e) {
-                    System.out.println("Thread Error: " + e.getMessage());
-                } finally {
-                    latch.countDown();
-                }
-            });
-        }
-
-        latch.await();
-        executor.shutdown();
-
-        System.out.println("시나리오1 테스트 완료");
-
-    }
-}
